+++ client/lib/main.dart
@@ -29,7 +29,9 @@ import 'package:flet_spinkit/flet_spinkit.dart' as flet_spinkit;
 // --FAT_CLIENT_START--
 import 'package:flet_video/flet_video.dart' as flet_video;
 // --FAT_CLIENT_END--
+// --WEBVIEW_IMPORT_START--
 import 'package:flet_webview/flet_webview.dart' as flet_webview;
+// --WEBVIEW_IMPORT_END--
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_web_plugins/url_strategy.dart';
@@ -63,7 +65,9 @@ void main([List<String>? args]) async {
     flet_permission_handler.Extension(),
     flet_secure_storage.Extension(),
     flet_spinkit.Extension(),
+    // --WEBVIEW_EXTENSION_START--
     flet_webview.Extension(),
+    // --WEBVIEW_EXTENSION_END--
 
     // --FAT_CLIENT_START--
     // --RIVE_EXTENSION_START--